feat: handle existing nextcloud user case

+++ synapse/handlers/watcha_partner.py
@@ -2,7 +2,7 @@ import logging
 
 from jsonschema.exceptions import SchemaError, ValidationError
 
-from synapse.api.errors import HttpResponseException, SynapseError
+from synapse.api.errors import HttpResponseException, NextcloudError
 from synapse.config.emailconfig import ThreepidBehaviour
 from synapse.push.mailer import Mailer
 from synapse.util.watcha import Secrets
@@ -48,9 +48,17 @@ class PartnerHandler(BaseHandler):
 
         try:
             await self.nextcloud_client.add_user(keycloak_user_id)
-        except (SynapseError, HttpResponseException, ValidationError, SchemaError):
-            await self.keycloak_client.delete_user(keycloak_user_id)
-            raise
+        except (
+            NextcloudError,
+            HttpResponseException,
+            ValidationError,
+            SchemaError,
+        ) as error:
+            if isinstance(error, NextcloudError) and error.code == 102:
+                logger.warn(f"[watcha] add user {keycloak_user_id} - failed : {error}")
+            else:
+                await self.keycloak_client.delete_user(keycloak_user_id)
+                raise
 
         invitee_id = await self.registration_handler.register_user(
             localpart=keycloak_user_id,
